move collapse to the main handler
