+++ numpyro/contrib/funsor/enum_messenger.py
@@ -9,17 +9,14 @@ from jax import lax
 import jax.numpy as jnp
 
 import funsor
-from funsor.terms import Funsor
-import numpyro
 from numpyro.handlers import trace as OrigTraceMessenger
-from numpyro.primitives import _PYRO_STACK, CondIndepStackFrame, Messenger, apply_stack
+from numpyro.primitives import CondIndepStackFrame, Messenger, apply_stack
 from numpyro.primitives import plate as OrigPlateMessenger
 
 funsor.set_backend("jax")
 
 
 __all__ = [
-    "collapse",
     "enum",
     "infer_config",
     "markov",
@@ -666,47 +663,3 @@ def to_data(x, name_to_dim=None, dim_type=DimType.LOCAL):
 
     msg = apply_stack(initial_msg)
     return msg['value']
-
-
-class collapse(OrigTraceMessenger):
-    def process_message(self, msg):
-        if msg["type"] == "sample":
-            if msg["value"] is None:
-                msg["value"] = msg["name"]
-
-            if isinstance(msg["fn"], Funsor) or isinstance(msg["value"], (str, Funsor)):
-                msg["stop"] = True
-
-    def __enter__(self):
-        self.preserved_plates = frozenset(h.name for h in _PYRO_STACK
-                                          if isinstance(h, OrigPlateMessenger))
-        return super().__enter__()
-
-    def __exit__(self, *args, **kwargs):
-        super().__exit__(*args, **kwargs)
-
-        # Convert delayed statements to pyro.factor()
-        reduced_vars = []
-        log_prob_terms = []
-        plates = frozenset()
-        for name, site in self.trace.items():
-            if not site["is_observed"]:
-                reduced_vars.append(name)
-            dim_to_name = {f.dim: f.name for f in site["cond_indep_stack"]}
-            fn = funsor.to_funsor(site["fn"], funsor.Real, dim_to_name)
-            value = site["value"]
-            if not isinstance(value, str):
-                value = funsor.to_funsor(site["value"], fn.inputs["value"], dim_to_name)
-            log_prob_terms.append(fn(value=value))
-            plates |= frozenset(f.name for f in site["cond_indep_stack"])
-        assert log_prob_terms, "nothing to collapse"
-        reduced_plates = plates - self.preserved_plates
-        log_prob = funsor.sum_product.sum_product(
-            funsor.ops.logaddexp,
-            funsor.ops.add,
-            log_prob_terms,
-            eliminate=frozenset(reduced_vars) | reduced_plates,
-            plates=plates,
-        )
-        name = reduced_vars[0]
-        numpyro.factor(name, log_prob.data)
